pre-commit/hooks/pyproject.toml
[project]
name = "coding-ethos-hooks"
version = "0.1.0"
description = "Pre-commit hooks for coding-ethos bundles"
requires-python = ">=3.13"
license = { text = "MIT" }

dependencies = [
    "black>=26.1.0",
    "google-api-core>=2.0.0",
    "google-genai>=1.0.0",
    "interrogate>=1.7.0",
    "mypy>=1.19.1",
    "mypy-extensions>=1.1.0",
    "pip>=26.0.1",
    "pydantic>=2.0.0",
    "pylint>=4.0.4",
    "pyright>=1.1.408",
    "pyupgrade>=3.21.2",
    "pyyaml>=6.0",
    "radon>=6.0.1",
    "ruff>=0.15.1",
    "types-aiofiles>=25.1.0.20260409",
    "types-beautifulsoup4>=4.12.0.20250516",
    "types-decorator>=5.2.0.20260408",
    "types-docker>=7.1.0.20260409",
    "types-openpyxl>=3.1.5.20260408",
    "types-pexpect>=4.9.0.20260408",
    "types-pillow>=10.2.0.20240822",
    "types-psutil>=7.2.2.20260408",
    "types-psycopg2>=2.9.21.20260422",
    "types-pyasn1>=0.6.0.20260408",
    "types-pycurl>=7.45.7.20260408",
    "types-pygments>=2.20.0.20260408",
    "types-pymysql>=1.1.0.20260408",
    "types-tqdm>=4.67.3.20260408",
    "types-ujson>=5.10.0.20250822",
    "types-xlrd>=2.0.0.20260408",
    "uv>=0.10.4",
    "vulture>=2.14",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["coding_ethos_hooks"]

[tool.pyright]
pythonVersion = "3.13"

[tool.mypy]
strict = true
python_version = "3.13"
plugins = ["pydantic.mypy"]

[tool.pydantic-mypy]
init_forbid_extra = true
init_typed = true
warn_required_dynamic_aliases = true

[tool.ruff]
target-version = "py313"
line-length = 88

[tool.ruff.lint]
select = ["ALL"]
ignore = [
    "COM812",
    "ISC001",
    "D203",
    "D213",
    "S603",
    "S607",
    "T201",
    "TC",
    "TC001",
    "TC002",
    "TC003",
    "TC004",
]

[tool.ruff.lint.pylint]
max-args = 6

[tool.ruff.lint.per-file-ignores]
"stubs/**" = [
    "A001",
    "A002",
    "ANN401",
    "D",
    "FBT001",
    "FBT002",
    "N801",
    "N803",
    "PYI020",
    "RUF100",
    "TCH",
    "UP037",
]
"**/stubs/**" = [
    "A001",
    "A002",
    "ANN401",
    "D",
    "FBT001",
    "FBT002",
    "N801",
    "N803",
    "PYI020",
    "RUF100",
    "TCH",
    "UP037",
]
"tests/**" = [
    "ANN401",
    "ARG001",
    "ARG002",
    "B017",
    "D",
    "E501",
    "EM101",
    "EM102",
    "PERF401",
    "PLR0913",
    "PLR2004",
    "PT011",
    "PT012",
    "RUF012",
    "S101",
    "S105",
    "S106",
    "SIM117",
    "SLF001",
    "TC003",
    "TRY003",
]
"**/tests/**" = [
    "ANN401",
    "ARG001",
    "ARG002",
    "B017",
    "D",
    "E501",
    "EM101",
    "EM102",
    "PERF401",
    "PLR0913",
    "PLR2004",
    "PT011",
    "PT012",
    "RUF012",
    "S101",
    "S105",
    "S106",
    "SIM117",
    "SLF001",
    "TC003",
    "TRY003",
]

[tool.ruff.lint.flake8-tidy-imports.banned-api]
"dateutil.tz" = { msg = "Use `zoneinfo` instead." }
"importlib" = { msg = "banned, use standard imports" }
